+++ Scently/Scently/API/Core/PerfumeTarget.swift
@@ -1,0 +1,8 @@
+//
+//  PerfumeTarget.swift
+//  Scently
+//
+//  Created by 임재현 on 7/13/25.
+//
+
+import Foundation
